coding sltp change logic to send bd connection to service

+++ domains/Student.py
@@ -1,7 +1,8 @@
 from pydantic import BaseModel
+from typing import List
 
 class StudentCreate(BaseModel):
   DNI: str
   nombres: str
-  notas:List[float]
-  promedio: int
+  notas:List[int]
+  promedio: float
